+++ src/backend/joanie/urls.py
@@ -26,6 +26,7 @@ from joanie.payment.urls import urlpatterns as payment_urlpatterns
 
 router = DefaultRouter()
 router.register("addresses", api.AddressViewSet, basename="addresses")
+router.register("certificates", api.CertificateViewSet, basename="certificates")
 router.register("courses", api.CourseViewSet, basename="courses")
 router.register("enrollments", api.EnrollmentViewSet, basename="enrollments")
 router.register("orders", api.OrderViewSet, basename="orders")
@@ -33,7 +34,6 @@ router.register("orders", api.OrderViewSet, basename="orders")
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/", include([*router.urls, *lms_urlpatterns, *payment_urlpatterns])),
-    path("api/documents/", include("marion.urls")),
 ]
 
 if settings.DEBUG: