+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased]
 
 ### Added
 
+- Add an api endpoint to retrieve and download certificates
 - Add a `get_grades` method to LMSHandler
 - Add a computed "state" to Course and CourseRun models
 - Add Payplug payment backend and related API routes